Separated course details tabs from the learn screen.
