+++ libs/horizon/src/main/java/com/instructure/horizon/features/home/HomeNavigation.kt
@@ -28,8 +28,8 @@ import androidx.navigation.navArgument
 import com.instructure.horizon.features.account.navigation.AccountNavigation
 import com.instructure.horizon.features.dashboard.DashboardScreen
 import com.instructure.horizon.features.dashboard.DashboardViewModel
-import com.instructure.horizon.features.learn.course.CourseDetailsScreen
-import com.instructure.horizon.features.learn.course.CourseDetailsViewModel
+import com.instructure.horizon.features.learn.LearnScreen
+import com.instructure.horizon.features.learn.LearnViewModel
 import com.instructure.horizon.features.skillspace.SkillspaceScreen
 import com.instructure.horizon.features.skillspace.SkillspaceViewModel
 import com.instructure.horizon.horizonui.animation.mainEnterTransition
@@ -73,9 +73,9 @@ fun HomeNavigation(navController: NavHostController, mainNavController: NavHostC
                     defaultValue = -1
                 }
             )) {
-            val viewModel = hiltViewModel<CourseDetailsViewModel>()
+            val viewModel = hiltViewModel<LearnViewModel>()
             val uiState by viewModel.state.collectAsState()
-            CourseDetailsScreen(uiState, mainNavController)
+            LearnScreen(uiState, mainNavController)
         }
         composable(HomeNavigationRoute.Skillspace.route) {
             val viewModel = hiltViewModel<SkillspaceViewModel>()
