+++ libs/horizon/src/main/java/com/instructure/horizon/features/learn/LearnUiState.kt
@@ -1,0 +1,26 @@
+/*
+ * Copyright (C) 2025 - present Instructure, Inc.
+ *
+ *     Licensed under the Apache License, Version 2.0 (the "License");
+ *     you may not use this file except in compliance with the License.
+ *     You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *     Unless required by applicable law or agreed to in writing, software
+ *     distributed under the License is distributed on an "AS IS" BASIS,
+ *     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *     See the License for the specific language governing permissions and
+ *     limitations under the License.
+ */
+package com.instructure.horizon.features.learn
+
+import com.instructure.canvasapi2.managers.CourseWithProgress
+import com.instructure.horizon.horizonui.platform.LoadingState
+
+data class LearnUiState(
+    val screenState: LoadingState = LoadingState(),
+    val courses: List<CourseWithProgress> = emptyList(),
+    val selectedCourse: CourseWithProgress? = null,
+    val onSelectedCourseChanged: ((CourseWithProgress) -> Unit) = {}
+)
